Syncing english and german contents
I've accidentally started with the english part of the reference which still includes german sentences.
There were several changes done in the past just in the english part of the reference, so that I don't know which one was the correct one.
Examples: 
Description of 'ASC_brightnessDriveUpDown', 'ASC_Drive_Offset' and 'ASC_WindProtection'

I hope that I found the correct ones.

---
FYI: Till now - only the english part is the primary part. I will swap the contents soon.