``` Add usage information for roller shutter control
This update enhances the documentation by adding crucial
usage information for the roller shutter control module.
New sections in both English and German were introduced
to explain the necessary conditions for preventing delayed
operation when a window contact is opened.

It is clarified that for the functionality to operate
as intended, the roller shutter must be open, the sensor
needs to be set to open, and specific attributes
(ASC_ShuttersPlace and ASC_Self_Defense_Mode) must
be configured correctly. These additions aim to improve
user understanding and usability of the AutoShuttersControl
module. The version was updated to v1.0.1 to reflect
these enhancements.
```